Move browser modules into WebView and wire them up
